[refactor] 각 API들을 파일을 만들어서 분리
